gather/requirements.txt
aniso8601==1.2.0
ansible==2.3.0.0
appdirs==1.4.3
asn1crypto==0.22.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.6
configparser==3.5.0
cryptography==1.8.1
decorator==4.0.11
ecdsa==0.13
Editra==0.7.20
entrypoints==0.2.2
enum34==1.1.6
Flask==0.11.1
Flask-Auth==0.85
Flask-Cors==3.0.2
Flask-RESTful==0.3.5
functools32==3.2.3.post2
idna==2.5
ipaddress==1.0.18
ipykernel==4.3.1
ipython==5.3.0
ipython-genutils==0.1.0
ipywidgets==5.1.5
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==5.0.0
jupyter-core==4.1.0
MarkupSafe==1.0
meld==3.16.4
mistune==0.7.3
mysql-connector-python==2.1.6
MySQL-python==1.2.5
nbconvert==4.2.0
nbformat==4.0.1
notebook==4.2.1
packaging==16.8
paramiko==2.1.2
path.py==8.2.1
pathlib==1.0.1
pexpect==4.2.1
pickleshare==0.7.4
ply==3.10
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.2.3
pycairo==1.13.1
pycparser==2.17
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.24.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==15.2.0
qtconsole==4.2.1
querycsv==0.3.0.0
rdflib==3.1.0
requests==2.18.1
selenium==3.4.3
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
tornado==4.3
traitlets==4.3.2
urllib3==1.21.1
virtualenv==15.0.3
wcwidth==0.1.7
Werkzeug==0.11.11
widgetsnbextension==1.2.3
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xmltodict==0.11.0
zenmap==7.40
